+++ src/SFA.DAS.Roatp.Data/Repositories/ProviderRegistrationDetailsWriteRepository.cs
@@ -4,9 +4,9 @@ using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
-using SFA.DAS.Roatp.Domain.Constants;
 using SFA.DAS.Roatp.Domain.Entities;
 using SFA.DAS.Roatp.Domain.Interfaces;
+using SFA.DAS.Roatp.Domain.Models;
 
 namespace SFA.DAS.Roatp.Data.Repositories
 {
@@ -23,9 +23,9 @@ namespace SFA.DAS.Roatp.Data.Repositories
         public async Task<List<ProviderRegistrationDetail>> GetActiveProviders() =>
             await _roatpDataContext.ProviderRegistrationDetails
             .Where(x =>
-                x.StatusId == OrganisationStatus.Active ||
-                x.StatusId == OrganisationStatus.ActiveNotTakingOnApprentices ||
-                x.StatusId == OrganisationStatus.Onboarding)
+                x.StatusId == (int)OrganisationStatus.Active ||
+                x.StatusId == (int)OrganisationStatus.ActiveNotTakingOnApprentices ||
+                x.StatusId == (int)OrganisationStatus.Onboarding)
             .ToListAsync();
 
         public async Task UpdateProviders(DateTime timeStarted, int providerCount, ImportType importType)
